This is our Progress Report
We had much difficulty implementing the app idea mentioned in the
proposal, mainly due to changing the images within Dietl's cannon game ,
which was the game that our app was to be based off of, to the pirate
sprites. Thus, we changed our idea to a game in which Bob's sprite  must
avoid obstacles (other pirates) and get the treasure. Right now, the
sprites won't load , so red squares represent the pirates, while green
represents the treasure. We also had trouble connecting the menu to the
app, so only the help and about buttons work. A red rectangle will
appear at the top of the screen if there is a collision with any of the
red squares and a green one will appear if there is a collision with the
green one, indicating loss and victory respectively. We hope to have at
least some of these issues fixed by the final report submission. This
submission contains the code and a mock-up for our app as it is now.